Server/allevents_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-elasticsearch-dsl==8.0
django-extensions==4.0
djangorestframework==3.16.0
drf-yasg==1.21.10
elastic-transport==8.17.1
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
musicbrainzngs==0.7.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
Wikipedia-API==0.8.1
